requirements.txt
# Black Ops Framework Requirements v3.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
colorama>=0.4.6
cryptography>=38.0.0
PyYAML>=6.0
pydantic>=2.8.0
scapy>=2.5.0
dnspython>=2.3.0
paramiko>=3.0.0
mysql-connector-python>=8.0.0
netifaces>=0.11.0
python-nmap>=0.7.1
stem>=1.8.2
PySocks>=1.7.1
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0
python-docx>=0.8.11
reportlab>=3.6.0
fpdf2>=2.7.8
flask>=3.0.0
selenium>=4.8.0
webdriver-manager>=3.8.0
tweepy>=4.12.0
openai>=1.0.0
playwright>=1.40.0
boto3>=1.34.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.10.0
progress>=1.6
tabulate>=0.9.0
python-magic>=0.4.27
psutil>=5.9.0
pyfiglet>=0.8.post1
watchdog>=3.0.0
pyjwt>=2.8.0
phonenumbers>=8.13.28
exif>=1.6.0
whois>=0.9.0
geoip2>=4.6.0
maxminddb>=2.0.0
pycryptodome>=3.19.0
hashid>=3.0.0
pytest>=7.2.0
pytest-cov>=4.0.0
